Add a way to exclude tests that rely on typed keys being true
