+++ src/Tests/Framework/XUnitPlumbing/NeedsTypedKeysAttribute.cs
@@ -1,0 +1,6 @@
+using System;
+
+namespace Tests.Framework
+{
+	public class NeedsTypedKeysAttribute : Attribute { }
+}
